add snackbars and loading indicators
